requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
mlxtend==0.19.0
paho-mqtt==1.5.1
hdfs==2.6.0
kafka-python==2.0.2
